기존의 MainActivity -> SearchPlaceActivity 지도가 나오는 액티비티를 MainActivity로 설정, main의 텍스트뷰를 누르면 검색화면으로 이동
