+++ app/src/main/AndroidManifest.xml
@@ -15,10 +15,10 @@
         android:theme="@style/Theme.Map"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".SearchPlaceActivity"
             android:exported="false" />
         <activity
-            android:name=".KakaoMapActivity"
+            android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
