+++ app/src/main/res/layout/activity_main.xml
@@ -1,54 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <EditText
-        android:id="@+id/search"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:layout_marginLeft="20sp"
-        android:layout_marginTop="15sp"
-        android:hint="검색어를 입력해 주세요"
-        app:layout_constraintHeight_percent="0.1"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintWidth_default="percent"
-        app:layout_constraintWidth_percent="0.8" />
-
-    <ImageView
-        android:id="@+id/close_icon"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/close_40px"
-        app:layout_constraintBottom_toBottomOf="@id/search"
-        app:layout_constraintLeft_toRightOf="@id/search"
-        app:layout_constraintTop_toTopOf="@id/search" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/saved_search"
+    <com.kakao.vectormap.MapView
+        android:id="@+id/map_view"
         android:layout_width="match_parent"
-        android:layout_height="40dp"
-        app:layout_constraintBottom_toTopOf="@id/recyclerView"
-        app:layout_constraintTop_toBottomOf="@id/search"
-        tools:listitem="@layout/item_saved_search" />
+        android:layout_height="match_parent" />
 
+    <EditText
+        android:id="@+id/main_search"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_margin="40dp"
+        android:hint="검색어를 입력해 주세요"
+        android:background="@android:color/white"
+        android:padding="15dp"
+        android:elevation="4dp"
+        android:layout_gravity="top" />
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/saved_search"
-        tools:listitem="@layout/item_recyclerview" />
-
-
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+</FrameLayout>
